backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "cryptography>=46.0.5",
    "fastapi>=0.131.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyjwt[crypto]>=2.9.0",
    "pydantic-settings>=2.13.1",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "anyio[asyncio]>=4.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0",
    "respx>=0.22",
    "ruff>=0.15.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["E402"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
